+++ Client/src/main/java/net/minecraft/client/entity/EntityOtherPlayerMP.java
@@ -57,8 +57,8 @@ public class EntityOtherPlayerMP extends AbstractClientPlayer
         this.otherPlayerMPX = x;
         this.otherPlayerMPY = y;
         this.otherPlayerMPZ = z;
-        this.otherPlayerMPYaw = (double)yaw;
-        this.otherPlayerMPPitch = (double)pitch;
+        this.otherPlayerMPYaw = yaw;
+        this.otherPlayerMPPitch = pitch;
         this.otherPlayerMPPosRotationIncrements = posRotationIncrements;
     }
 
@@ -98,7 +98,6 @@ public class EntityOtherPlayerMP extends AbstractClientPlayer
 
             for (d3 = this.otherPlayerMPYaw - (double)this.rotationYaw; d3 < -180.0D; d3 += 360.0D)
             {
-                ;
             }
 
             while (d3 >= 180.0D)
